backend/requirements.txt
django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
load_dotenv
certifi
gunicorn